Surface keybindings in server config updates
- Include resolved keybindings in the websocket payload
- Share keybinding compilation logic and defaults in `packages/shared`
- Update client state and tests to use the streamed keybindings
